requirements.txt
yfinance
pandas
numpy
scikit-learn
joblib
ta
streamlit
confluent-kafka
python-dateutil
